+++ api/response/ResponseTypes.ts
@@ -1,4 +1,4 @@
-import { EvaluatedProduct, KassalappProduct } from "./ProductTypes"
+import { EvaluatedProduct, KassalappProduct } from "../../types/ProductTypes"
 
 export default interface KassalappResponse {
     data: {
